@@ -5,14 +5,14 @@ build-backend = "pdm.backend"
 [project]
 authors = [{ name = "Erick Friis", email = "erick@langchain.dev" }]
 license = { text = "MIT" }
-requires-python = ">=3.9"
+requires-python = ">=3.10.0,<4.0.0"
 dependencies = [
-    "typer<1.0.0,>=0.9.0",
-    "gitpython<4,>=3",
-    "langserve[all]>=0.0.51",
-    "uvicorn<1.0,>=0.23",
-    "tomlkit>=0.12",
-    "gritql<1.0.0,>=0.2.0",
+    "typer>=0.17.0,<1.0.0",
+    "gitpython>=3,<4.0.0",
+    "langserve[all]>=0.0.51,<1.0.0",
+    "uvicorn>=0.23.0,<1.0.0",
+    "tomlkit>=0.12.0,<1.0.0",
+    "gritql>=0.2.0,<1.0.0",
 ]
 name = "langchain-cli"
 version = "0.0.37"
@@ -29,8 +29,8 @@ langchain = "langchain_cli.cli:app"
 langchain-cli = "langchain_cli.cli:app"
 
 [dependency-groups]
-dev = ["pytest<9.0.0,>=7.4.2", "pytest-watcher<1.0.0,>=0.3.4"]
-lint = ["ruff<0.13,>=0.12.2", "mypy<1.19,>=1.18.1"]
+dev = ["pytest>=7.4.2,<9.0.0", "pytest-watcher>=0.3.4,<1.0.0"]
+lint = ["ruff>=0.12.2,<0.13", "mypy>=1.18.1,<1.19"]
 test = ["langchain-core", "langchain"]
 typing = ["langchain"]
 test_integration = []
@@ -38,9 +38,6 @@ test_integration = []
 [tool.uv.sources]
 langchain-core = { path = "../core", editable = true }
 langchain = { path = "../langchain", editable = true }
-
-[tool.ruff]
-target-version = "py39"
 
 [tool.ruff.format]
 docstring-code-format = true
@@ -79,6 +76,12 @@ pyupgrade.keep-runtime-typing = true
 [tool.ruff.lint.per-file-ignores]
 "tests/**" = [ "D1", "DOC", "S", "SLF",]
 "scripts/**" = [ "INP", "S",]
+
+[tool.pytest.ini_options]
+addopts = "--strict-markers --strict-config --durations=5"
+markers = [
+    "compile: mark placeholder test used to compile integration tests without running them",
+]
 
 [tool.mypy]
 plugins = ["pydantic.mypy"]
