@@ -45,7 +45,6 @@ class InputTokenDetails(TypedDict, total=False):
     Does *not* need to sum to full input token count. Does *not* need to have all keys.
 
     Example:
-
         .. code-block:: python
 
             {
@@ -72,6 +71,7 @@ class InputTokenDetails(TypedDict, total=False):
 
     Since there was a cache hit, the tokens were read from the cache. More precisely,
     the model state given these tokens was read from the cache.
+
     """
 
 
@@ -81,7 +81,6 @@ class OutputTokenDetails(TypedDict, total=False):
     Does *not* need to sum to full output token count. Does *not* need to have all keys.
 
     Example:
-
         .. code-block:: python
 
             {
@@ -100,6 +99,7 @@ class OutputTokenDetails(TypedDict, total=False):
 
     Tokens generated by the model in a chain of thought process (i.e. by OpenAI's o1
     models) that are not returned as part of model output.
+
     """
 
 
@@ -109,7 +109,6 @@ class UsageMetadata(TypedDict):
     This is a standard representation of token usage that is consistent across models.
 
     Example:
-
         .. code-block:: python
 
             {
@@ -148,6 +147,7 @@ class UsageMetadata(TypedDict):
     """Breakdown of output token counts.
 
     Does *not* need to sum to full output token count. Does *not* need to have all keys.
+
     """
 
 
@@ -159,12 +159,14 @@ class AIMessage(BaseMessage):
     This message represents the output of the model and consists of both
     the raw output as returned by the model together standardized fields
     (e.g., tool calls, usage metadata) added by the LangChain framework.
+
     """
 
     example: bool = False
     """Use to denote that a message is part of an example conversation.
 
     At the moment, this is ignored by most models. Usage is discouraged.
+
     """
 
     tool_calls: list[ToolCall] = []
@@ -175,6 +177,7 @@ class AIMessage(BaseMessage):
     """If provided, usage metadata for a message, such as token counts.
 
     This is a standard representation of token usage that is consistent across models.
+
     """
     raw_response: Optional[Any] = Field(default=None, exclude=True)
     """Raw response object from the model provider.
@@ -195,12 +198,14 @@ class AIMessage(BaseMessage):
     """
 
     type: Literal["ai"] = "ai"
-    """The type of the message (used for deserialization). Defaults to "ai"."""
+    """The type of the message (used for deserialization). Defaults to ``'ai'``."""
 
     def __init__(
-        self, content: Union[str, list[Union[str, dict]]], **kwargs: Any
+        self,
+        content: Union[str, list[Union[str, dict]]],
+        **kwargs: Any,
     ) -> None:
-        """Pass in content as positional arg.
+        """Initialize ``AIMessage``.
 
         Args:
             content: The content of the message.
@@ -271,6 +276,7 @@ class AIMessage(BaseMessage):
 
         Returns:
             A pretty representation of the message.
+
         """
         base = super().pretty_repr(html=html)
         lines = []
@@ -310,7 +316,10 @@ class AIMessageChunk(AIMessage, BaseMessageChunk):
     # non-chunk variant.
     type: Literal["AIMessageChunk"] = "AIMessageChunk"  # type: ignore[assignment]
     """The type of the message (used for deserialization).
-    Defaults to "AIMessageChunk"."""
+
+    Defaults to ``AIMessageChunk``.
+
+    """
 
     tool_call_chunks: list[ToolCallChunk] = []
     """If provided, tool call chunks associated with the message."""
@@ -328,7 +337,10 @@ class AIMessageChunk(AIMessage, BaseMessageChunk):
         """Initialize tool calls from tool call chunks.
 
         Returns:
-            This ``AIMessageChunk``.
+            The values with tool calls initialized.
+
+        Raises:
+            ValueError: If the tool call chunks are malformed.
         """
         if not self.tool_call_chunks:
             if self.tool_calls:
@@ -553,9 +565,9 @@ def add_usage(
 def subtract_usage(
     left: Optional[UsageMetadata], right: Optional[UsageMetadata]
 ) -> UsageMetadata:
-    """Recursively subtract two UsageMetadata objects.
+    """Recursively subtract two ``UsageMetadata`` objects.
 
-    Token counts cannot be negative so the actual operation is max(left - right, 0).
+    Token counts cannot be negative so the actual operation is ``max(left - right, 0)``.
 
     Example:
         .. code-block:: python
