@@ -1,7 +1,5 @@
 """Configuration file for the Sphinx documentation builder."""
 
-# Configuration file for the Sphinx documentation builder.
-#
 # This file only contains a selection of the most common options. For a full
 # list see the documentation:
 # https://www.sphinx-doc.org/en/master/usage/configuration.html
@@ -20,16 +18,18 @@ from docutils.parsers.rst.directives.admonitions import BaseAdmonition
 from docutils.statemachine import StringList
 from sphinx.util.docutils import SphinxDirective
 
-# If extensions (or modules to document with autodoc) are in another directory,
-# add these directories to sys.path here. If the directory is relative to the
-# documentation root, use os.path.abspath to make it absolute, like shown here.
-
+# Add paths to Python import system so Sphinx can import LangChain modules
+# This allows autodoc to introspect and document the actual code
 _DIR = Path(__file__).parent.absolute()
-sys.path.insert(0, os.path.abspath("."))
-sys.path.insert(0, os.path.abspath("../../libs/langchain"))
+sys.path.insert(0, os.path.abspath("."))  # Current directory
+sys.path.insert(0, os.path.abspath("../../libs/langchain"))  # LangChain main package
 
+# Load package metadata from pyproject.toml (for version info, etc.)
 with (_DIR.parents[1] / "libs" / "langchain" / "pyproject.toml").open("r") as f:
     data = toml.load(f)
+
+# Load mapping of classes to example notebooks for backlinking
+# This file is generated by scripts that scan our tutorial/example notebooks
 with (_DIR / "guide_imports.json").open("r") as f:
     imported_classes = json.load(f)
 
@@ -86,6 +86,7 @@ class Beta(BaseAdmonition):
 
 
 def setup(app):
+    """Register custom directives and hooks with Sphinx."""
     app.add_directive("example_links", ExampleLinksDirective)
     app.add_directive("beta", Beta)
     app.connect("autodoc-skip-member", skip_private_members)
@@ -258,7 +259,20 @@ html_static_path = ["_static"]
 html_css_files = ["css/custom.css"]
 html_use_index = False
 
-myst_enable_extensions = ["colon_fence"]
+myst_enable_extensions = [
+    "colon_fence",  # ::: directive blocks (existing prior to LangGraph support)
+    # LangGraph compatibility extensions added for consolidation
+    # TODO: check for presence of each in LangGraph and only enable if needed
+    # "deflist",  # Definition lists
+    # "tasklist",  # - [ ] checkboxes (common in examples)
+    # "attrs_inline",  # {.class} inline attributes (MkDocs style)
+    # "attrs_block",  # Block-level attributes
+    # "substitution",  # Variable substitution
+    # "linkify",  # Auto-link URLs in text
+    # Math extensions (uncomment if LangGraph uses mathematical notation)
+    # "dollarmath",     # $ math $ inline math
+    # "amsmath",        # Advanced math environments
+]
 
 # generate autosummary even if no references
 autosummary_generate = True
@@ -268,10 +282,10 @@ autosummary_ignore_module_all = False
 html_copy_source = False
 html_show_sourcelink = False
 
+googleanalytics_id = "G-9B66JQQH2F"
+
 # Set canonical URL from the Read the Docs Domain
 html_baseurl = os.environ.get("READTHEDOCS_CANONICAL_URL", "")
-
-googleanalytics_id = "G-9B66JQQH2F"
 
 # Tell Jinja2 templates the build is running on Read the Docs
 if os.environ.get("READTHEDOCS", "") == "True":
