@@ -11,7 +11,7 @@ class TestParrotMultiplyToolUnit(ToolsUnitTests):
 
     @property
     def tool_constructor_params(self) -> dict:
-        # if your tool constructor instead required initialization arguments like
+        # If your tool constructor instead required initialization arguments like
         # `def __init__(self, some_arg: int):`, you would return those here
         # as a dictionary, e.g.: `return {'some_arg': 42}`
         return {}
@@ -21,7 +21,7 @@ class TestParrotMultiplyToolUnit(ToolsUnitTests):
         """
         Returns a dictionary representing the "args" of an example tool call.
 
-        This should NOT be a ToolCall dict - i.e. it should not
-        have {"name", "id", "args"} keys.
+        This should NOT be a ToolCall dict - i.e. it should not have
+        `{"name", "id", "args"}` keys.
         """
         return {"a": 2, "b": 3}
