@@ -65,33 +65,43 @@ class BaseChatMessageHistory(ABC):
 
         .. code-block:: python
 
+            import json
+            import os
+            from langchain_core.messages import messages_from_dict, message_to_dict
+
+
             class FileChatMessageHistory(BaseChatMessageHistory):
                 storage_path: str
                 session_id: str
 
                 @property
-                def messages(self):
-                    with open(
-                        os.path.join(storage_path, session_id),
-                        "r",
-                        encoding="utf-8",
-                    ) as f:
-                        messages = json.loads(f.read())
-                    return messages_from_dict(messages)
+                def messages(self) -> list[BaseMessage]:
+                    try:
+                        with open(
+                            os.path.join(self.storage_path, self.session_id),
+                            "r",
+                            encoding="utf-8",
+                        ) as f:
+                            messages_data = json.load(f)
+                        return messages_from_dict(messages_data)
+                    except FileNotFoundError:
+                        return []
 
                 def add_messages(self, messages: Sequence[BaseMessage]) -> None:
                     all_messages = list(self.messages)  # Existing messages
                     all_messages.extend(messages)  # Add new messages
 
                     serialized = [message_to_dict(message) for message in all_messages]
-                    # Can be further optimized by only writing new messages
-                    # using append mode.
-                    with open(os.path.join(storage_path, session_id), "w") as f:
-                        json.dump(messages, f)
+                    file_path = os.path.join(self.storage_path, self.session_id)
+                    os.makedirs(os.path.dirname(file_path), exist_ok=True)
+                    with open(file_path, "w", encoding="utf-8") as f:
+                        json.dump(serialized, f)
 
-                def clear(self):
-                    with open(os.path.join(storage_path, session_id), "w") as f:
-                        f.write("[]")
+                def clear(self) -> None:
+                    file_path = os.path.join(self.storage_path, self.session_id)
+                    os.makedirs(os.path.dirname(file_path), exist_ok=True)
+                    with open(file_path, "w", encoding="utf-8") as f:
+                        json.dump([], f)
 
     """
 
