@@ -15,19 +15,20 @@ from langchain_core.utils._merge import merge_dicts
 class FunctionMessage(BaseMessage):
     """Message for passing the result of executing a tool back to a model.
 
-    FunctionMessage are an older version of the ToolMessage schema, and
-    do not contain the tool_call_id field.
+    ``FunctionMessage`` are an older version of the ``ToolMessage`` schema, and
+    do not contain the ``tool_call_id`` field.
 
-    The tool_call_id field is used to associate the tool call request with the
+    The ``tool_call_id`` field is used to associate the tool call request with the
     tool call response. This is useful in situations where a chat model is able
     to request multiple tool calls in parallel.
+
     """
 
     name: str
     """The name of the function that was executed."""
 
     type: Literal["function"] = "function"
-    """The type of the message (used for serialization). Defaults to "function"."""
+    """The type of the message (used for serialization). Defaults to ``'function'``."""
 
 
 class FunctionMessageChunk(FunctionMessage, BaseMessageChunk):
@@ -38,7 +39,10 @@ class FunctionMessageChunk(FunctionMessage, BaseMessageChunk):
     # non-chunk variant.
     type: Literal["FunctionMessageChunk"] = "FunctionMessageChunk"  # type: ignore[assignment]
     """The type of the message (used for serialization).
-    Defaults to "FunctionMessageChunk"."""
+
+    Defaults to ``'FunctionMessageChunk'``.
+
+    """
 
     @override
     def __add__(self, other: Any) -> BaseMessageChunk:  # type: ignore[override]
