@@ -28,7 +28,11 @@ class SystemMessage(BaseMessage):
     """
 
     type: Literal["system"] = "system"
-    """The type of the message (used for serialization). Defaults to "system"."""
+    """The type of the message (used for serialization).
+
+    Defaults to ``'system'``.
+
+    """
 
     def __init__(
         self, content: Union[str, list[Union[str, dict]]], **kwargs: Any
@@ -50,4 +54,7 @@ class SystemMessageChunk(SystemMessage, BaseMessageChunk):
     # non-chunk variant.
     type: Literal["SystemMessageChunk"] = "SystemMessageChunk"  # type: ignore[assignment]
     """The type of the message (used for serialization).
-    Defaults to "SystemMessageChunk"."""
+
+    Defaults to ``'SystemMessageChunk'``.
+
+    """
