@@ -15,34 +15,35 @@ class AnyStr(str):
 
 # The code below creates version of pydantic models
 # that will work in unit tests with AnyStr as id field
+
 # Please note that the `id` field is assigned AFTER the model is created
 # to workaround an issue with pydantic ignoring the __eq__ method on
 # subclassed strings.
 
 
 def _any_id_document(**kwargs: Any) -> Document:
-    """Create a document with an id field."""
+    """Create a `Document` with an id field."""
     message = Document(**kwargs)
     message.id = AnyStr()
     return message
 
 
 def _any_id_ai_message(**kwargs: Any) -> AIMessage:
-    """Create ai message with an any id field."""
+    """Create an `AIMessage` with an any id field."""
     message = AIMessage(**kwargs)
     message.id = AnyStr()
     return message
 
 
 def _any_id_ai_message_chunk(**kwargs: Any) -> AIMessageChunk:
-    """Create ai message with an any id field."""
+    """Create an `AIMessageChunk` with an any id field."""
     message = AIMessageChunk(**kwargs)
     message.id = AnyStr()
     return message
 
 
 def _any_id_human_message(**kwargs: Any) -> HumanMessage:
-    """Create a human with an any id field."""
+    """Create a `HumanMessage` with an any id field."""
     message = HumanMessage(**kwargs)
     message.id = AnyStr()
     return message
