@@ -28,12 +28,12 @@ class HuggingFaceEmbeddings(BaseModel, Embeddings):
             from langchain_huggingface import HuggingFaceEmbeddings
 
             model_name = "sentence-transformers/all-mpnet-base-v2"
-            model_kwargs = {'device': 'cpu'}
-            encode_kwargs = {'normalize_embeddings': False}
+            model_kwargs = {"device": "cpu"}
+            encode_kwargs = {"normalize_embeddings": False}
             hf = HuggingFaceEmbeddings(
                 model_name=model_name,
                 model_kwargs=model_kwargs,
-                encode_kwargs=encode_kwargs
+                encode_kwargs=encode_kwargs,
             )
 
     """
