@@ -61,7 +61,10 @@ class HuggingFacePipeline(BaseLLM):
             tokenizer = AutoTokenizer.from_pretrained(model_id)
             model = AutoModelForCausalLM.from_pretrained(model_id)
             pipe = pipeline(
-                "text-generation", model=model, tokenizer=tokenizer, max_new_tokens=10
+                "text-generation",
+                model=model,
+                tokenizer=tokenizer,
+                max_new_tokens=10,
             )
             hf = HuggingFacePipeline(pipeline=pipe)
 
