@@ -40,8 +40,8 @@ def test_initialize_more() -> None:
     assert llm.temperature == 0
 
     ls_params = llm._get_ls_params()
-    assert ls_params["ls_provider"] == "azure"
-    assert ls_params["ls_model_name"] == "gpt-35-turbo-0125"
+    assert ls_params.get("ls_provider") == "azure"
+    assert ls_params.get("ls_model_name") == "gpt-35-turbo-0125"
 
 
 def test_initialize_azure_openai_with_openai_api_base_set() -> None:
@@ -139,3 +139,37 @@ def test_chat_completions_api_uses_model_name() -> None:
     assert payload["model"] == "gpt-5"
     assert "messages" in payload  # Chat Completions API uses 'messages'
     assert "input" not in payload
+
+
+def test_max_completion_tokens_parameter() -> None:
+    """Test that max_completion_tokens can be used as a direct parameter."""
+    llm = AzureChatOpenAI(
+        azure_deployment="gpt-5",
+        api_version="2024-12-01-preview",
+        azure_endpoint="my-base-url",
+        max_completion_tokens=1500,
+    )
+    messages = [HumanMessage("Hello")]
+    payload = llm._get_request_payload(messages)
+
+    # Should use max_completion_tokens instead of max_tokens
+    assert "max_completion_tokens" in payload
+    assert payload["max_completion_tokens"] == 1500
+    assert "max_tokens" not in payload
+
+
+def test_max_tokens_converted_to_max_completion_tokens() -> None:
+    """Test that max_tokens is converted to max_completion_tokens."""
+    llm = AzureChatOpenAI(
+        azure_deployment="gpt-5",
+        api_version="2024-12-01-preview",
+        azure_endpoint="my-base-url",
+        max_tokens=1000,  # type: ignore[call-arg]
+    )
+    messages = [HumanMessage("Hello")]
+    payload = llm._get_request_payload(messages)
+
+    # max_tokens should be converted to max_completion_tokens
+    assert "max_completion_tokens" in payload
+    assert payload["max_completion_tokens"] == 1000
+    assert "max_tokens" not in payload
